added all keypoints case

+++ db/db_access.py
@@ -146,8 +146,10 @@ def get_training_data_as_numpy(sqcur):
     )
     rows = sqcur.fetchall()
 
-    X = np.stack([np.asarray(create_image_from_pixels(row[0])) for row in rows])
-    y = np.asarray([row[1:] for row in rows])
+    X = np.stack([np.asarray(create_image_from_pixels(row[0])) for row in rows]).astype(
+        "float32"
+    )
+    y = np.asarray([row[1:] for row in rows]).astype("float32")
 
     print(f"Image array shape: {X.shape}, Label array shape: {y.shape}")
